fixed some bugs in the code

+++ src/game/Game.h
@@ -50,11 +50,16 @@ typedef struct GameObjectManager
 	size_t freeSpace;
 } GameObjectManager;
 
+
+// game manager struct??
+// contains all managers, e.g game object manager etc
+// not needed for this unit tbh
+
 void initGameObjectManager(GameObjectManager *gameObjectManager);
 void gameObjectManagerIncrease(GameObjectManager *gameObjectManager);
 void gameObjectManagerAdd(GameObjectManager *gameObjectManager, GameObject GameObject);
-void gameObjectManagerRemove(GameObjectManager *gameObjectManager, uint32_t id);
-GameObject* gameObjectManagerFind(GameObjectManager *gameObjectManager, uint32_t id);
+void gameObjectManagerRemove(GameObjectManager *gameObjectManager, size_t id);
+GameObject* gameObjectManagerFind(GameObjectManager *gameObjectManager, size_t id);
 
 void initGameObject(GameObject *gameObject); // initialise the game object
 void updateGameObject(GameObject *gameObject); // called every frame update
